+++ src/commands/vite/m.ts
@@ -11,16 +11,19 @@ const command: GluegunCommand = {
     if (filesystem.isDirectory(installedPath)) {
       return print.error(`Folder with the name [${appName}] already exists`)
     }
+
     const spinner = print.spin(`Creating vite app with react swc and tailwind [${appName}]`)
     await system.run(`bun create vite ${appName} --template react-swc-ts`)
     spinner.info(`Created vite app [${appName}]!`)
     process.chdir(installedPath)
+
     spinner.info(`Installing clsx and prettier!`)
     await system.run('bun i && ni clsx && fii add prettier && fii add tailwind && fii add release')
-    const gitInit = await system.run(`git flow init -d && git add . && git commit -m init`)
+
+    const { gitFlowInit } = await import('../../lib/helpers/git-utils')
+    const gitInit = await gitFlowInit()
     spinner.succeed(gitInit)
 
-    // spinner.info('Adding prettier and tailwind');
     spinner.succeed(`Created App ${appName}`)
     print.info(`webstorm ${installedPath}`)
     await system.run(`webstorm ${installedPath}`)