+++ src/extensions/cli-extension.ts
@@ -1,4 +1,3 @@
-import { BrowserCookiesSingleton } from '@ax-sh/browser-cookies'
 import { filesystem } from 'gluegun'
 
 import { type ExtendedToolbox, KnownError, UsableBinaryNotFound } from '../types'
@@ -7,6 +6,11 @@ module.exports = async (toolbox: ExtendedToolbox) => {
   const hasPnpm = toolbox.system.which('pnpm')
   const hasNi = toolbox.system.which('ni')
   const hasNr = toolbox.system.which('nr')
+  const hasPython = toolbox.system.which('python')
+  if (hasPython) {
+    process.env.PYTHON_BIN = hasPython
+  }
+
   if (!hasPnpm) {
     throw new Error('No pnpm available')
   }
@@ -29,7 +33,6 @@ module.exports = async (toolbox: ExtendedToolbox) => {
     const hasScript = script !== '{}'
     if (hasScript) {
       toolbox.print.error(`script ${scriptName} already defined Exiting`)
-      // print.spinner.stopAndPersist({ symbol: '🚨', text: '!' });
 
       throw new KnownError(`script ${scriptName} already defined Exiting`)
     }
@@ -37,11 +40,9 @@ module.exports = async (toolbox: ExtendedToolbox) => {
     await toolbox.system.run(`pnpm pkg set scripts.${scriptName}="${cmd}"`)
   }
 
-  const pythonPath = 'C:/Users/USER/miniconda3/python.exe'
-  process.env.PYTHON_BIN = pythonPath
-
-  // const { BrowserCookiesSingleton } = await import('@ax-sh/browser-cookies')
-  toolbox.loadBrowser = BrowserCookiesSingleton.instance
+  // const { BrowserCookiesSingleton } = await import('@ax-sh/browser-cookies/dist/browser-cookies.js')
+  // console.log(BrowserCookiesSingleton)
+  // toolbox.loadBrowser = BrowserCookiesSingleton.instance
 
   // enable this if you want to read configuration in from
   // the current folder's package.json (in a "fii" property),